MetaICVI

Documentation for MetaICVI.

These pages serve to summarize the internals and functionality of the Meta-ICVI approach. This project was developed by the Applied Computational Intelligence Laboratory (ACIL) of the Missouri University of Science and Technology (S&T) under the supervision of Teledyne Technologies for the DARPA L2M program.

Index

MetaICVI.MetaICVIModule

MetaICVI.MetaICVIModule

MetaICVI.MetaICVIModule

MetaICVI.MetaICVIOpts

MetaICVI.get_correlations

MetaICVI.get_icvis

MetaICVI.get_metaicvi

MetaICVI.get_rocket_features

Documentation

MetaICVI.MetaICVIModule — Type

MetaICVI

Fields

opts::MetaICVIOpts: options for construction.

cvis::Vector{AbstractCVI}: list of cvis used for computing the CVIs.

criterion_values::RealVector: list of outputs of the cvis used for computing correlations.

correlations::RealVector: list of outputs of the rank correlations.

features::RealVector: list of outputs of the rocket feature kernels.

rocket::RocketModule: time-series random feature kernels module.

performance::RealFP: final output of the most recent the Meta-ICVI step.

MetaICVI.MetaICVIModule — Method

MetaICVIModule(opts::MetaICVIOpts)

Instantiate a MetaICVIModule with given options.

Arguments

opts::MetaICVIOpts: options struct for the MetaICVI object.

MetaICVI.MetaICVIModule — Method

MetaICVIModule()

Default constructor for the MetaICVIModule.

MetaICVI.MetaICVIOpts — Type

MetaICVIOpts()

Meta-ICVI module options.

Examples

julia> MetaICVIOpts()

MetaICVI.get_correlations — Method

get_correlations(metaicvi::MetaICVIModule)

Compute and store the rank correlations from the cvi values.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

MetaICVI.get_icvis — Method

get_icvis(metaicvi::MetaICVIModule, sample::RealVector, label::Integer)

Compute and store the icvi criterion values.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

sample::RealVector: the sample used for clustering.

label::Integer: the label prescribed to the sample by the clustering algorithm.

MetaICVI.get_metaicvi — Method

get_metaicvi(metaicvi::MetaICVIModule, sample::RealVector, label::Integer)

Compute and return the meta-icvi value.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

sample::RealVector: the sample used for clustering.

label::Integer: the label prescribed to the sample by the clustering algorithm.

MetaICVI.get_rocket_features — Method

get_rocket_features(metaicvi::MetaICVIModule)

Compute and store the rocket features.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.