MetaICVI

Documentation for MetaICVI.

These pages serve to summarize the internals and functionality of the Meta-ICVI approach. This project was developed by the Applied Computational Intelligence Laboratory (ACIL) of the Missouri University of Science and Technology (S&T) under the supervision of Teledyne Technologies for the DARPA L2M program.

Index

MetaICVI.MetaICVIModule

MetaICVI.MetaICVIOpts

Documentation

MetaICVI.MetaICVIModule — Type

MetaICVI

MetaICVI.MetaICVIOpts — Type

MetaICVIOpts()

Meta-ICVI module options

Examples

julia> MetaICVIOpts()