Index

All structures and methods can be found in the Full Index with their accompanying docstrings in Documentation

Full Index

MetaICVI.MetaICVIModule

MetaICVI.MetaICVIModule

MetaICVI.MetaICVIModule

MetaICVI.MetaICVIOpts

MetaICVI.Rocket.RocketKernel

MetaICVI.Rocket.RocketModule

MetaICVI.Rocket.RocketModule

MetaICVI.Rocket.RocketModule

Base.show

MetaICVI.Rocket.apply_kernel

MetaICVI.Rocket.apply_kernels

MetaICVI.Rocket.load_rocket

MetaICVI.Rocket.save_rocket

MetaICVI.construct_classifier

MetaICVI.get_correlations

MetaICVI.get_cvi_data

MetaICVI.get_features

MetaICVI.get_icvis

MetaICVI.get_metaicvi

MetaICVI.get_probability

MetaICVI.get_rocket_features

MetaICVI.get_training_features

MetaICVI.is_pretrained

MetaICVI.load_classifier

MetaICVI.safe_save_classifier

MetaICVI.safe_save_rocket

MetaICVI.save_classifier

MetaICVI.train_and_save

Documentation

MetaICVI.MetaICVIModule — Type

MetaICVIModule

Stateful information for a single MetaICVI module.

Fields

opts::MetaICVIOpts: options for construction.

cvis::Vector{CVI}: list of cvis used for computing the CVIs.

criterion_values::RealVector: list of outputs of the cvis used for computing correlations.

correlations::RealVector: list of outputs of the rank correlations.

features::RealVector: list of outputs of the rocket feature kernels.

rocket::RocketModule: time-series random feature kernels module.

classifier::MetaICVIClassifier: ScikitLearn classifier.

performance::RealFP: final output of the most recent the Meta-ICVI step.

is_pretrained::Bool: internal flag for if the classifier is trained and ready for inference.

MetaICVI.MetaICVIModule — Method

MetaICVIModule(opts::MetaICVIOpts)

Instantiate a MetaICVIModule with given options.

Arguments

opts::MetaICVIOpts: options struct for the MetaICVI object.

MetaICVI.MetaICVIModule — Method

MetaICVIModule()

Default constructor for the MetaICVIModule.

MetaICVI.MetaICVIOpts — Type

MetaICVIOpts()

Meta-ICVI module options.

Examples

julia> MetaICVIOpts()

Base.show — Method

Base.show(io::IO, metaicvi::MetaICVIModule)

Display a metaicvi module to the command line.

Arguments

io::IO: default io stream.

metaicvi::MetaICVIModule: metaicvi object about which to display info.

MetaICVI.construct_classifier — Method

construct_classifier(opts::MetaICVIOpts)

Construct a new classifier for the MetaICVI module with metaprogramming.

Arguments

opts::MetaICVIOpts: options containing the classifier type and options for instantiation.

MetaICVI.get_correlations — Method

get_correlations(metaicvi::MetaICVIModule)

Compute and store the rank correlations from the cvi values.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

MetaICVI.get_cvi_data — Method

get_cvi_data(data_file::String)

Get the cvi data specified by the data_file path.

Arguments

data_file::String: file containing clustered data for cvi processing.

MetaICVI.get_features — Method

get_features(metaicvi::MetaICVIModule, sample::RealVector, label::Integer)

Compute only the features on the sample and label without classifier inference.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

sample::RealVector: the sample used for clustering.

label::Integer: the label prescribed to the sample by the clustering algorithm.

MetaICVI.get_icvis — Method

get_icvis(metaicvi::MetaICVIModule, sample::RealVector, label::Integer)

Compute and store the icvi criterion values.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

sample::RealVector: the sample used for clustering.

label::Integer: the label prescribed to the sample by the clustering algorithm.

MetaICVI.get_metaicvi — Method

get_metaicvi(metaicvi::MetaICVIModule, sample::RealVector, label::Integer)

Compute and return the meta-icvi value.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

sample::RealVector: the sample used for clustering.

label::Integer: the label prescribed to the sample by the clustering algorithm.

MetaICVI.get_probability — Method

get_probability(metaicvi::MetaICVIModule)

Compute and store the metaicvi value from the classifier.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

MetaICVI.get_rocket_features — Method

get_rocket_features(metaicvi::MetaICVIModule)

Compute and store the rocket features.

Arguments

metaicvi::MetaICVIModule: the Meta-ICVI module.

MetaICVI.get_training_features — Method

get_training_features(metaicvi::MetaICVIModule, data_path::String)

Arguments

MetaICVI.is_pretrained — Method

is_pretrained(metaicvi::MetaICVIModule)

Checks if the classifier is pretrained to permit inference.

Arguments

metaicvi::MetaICVIModule: metaicvi module containing the classifier to check.

MetaICVI.load_classifier — Method

load_classifier(filepath::String)

Load the classifier at the filepath.

Arguments

filepath::String: location of the classifier .jld file.

MetaICVI.safe_save_classifier — Method

safe_save_classifier(metaicvi::MetaICVIModule)

Error handle saving of the metaicvi classifier.

Arguments

metaicvi::MetaICVIModule: metaicvi module containing the classifier and path for saving.

MetaICVI.safe_save_rocket — Method

safe_save_rocket(metaicvi::MetaICVIModule)

Error handle saving of the metaicvi rocket kernels.

Arguments

metaicvi::MetaICVIModule: metaicvi module containing the classifier and path for saving.

MetaICVI.save_classifier — Method

save_classifier(classifier::MetaICVIClassifier, filepath::String)

Save the classifier at the filepath.

Arguments

classifier::MetaICVIClassifier: classifier object to save.

filepath::String: name/path to save the classifier .jld file.

MetaICVI.train_and_save — Method

train_and_save(metaicvi::MetaICVIModule, x::RealMatrix, y::IntegerVector)

Train the classifier on x/y and save the kernels and classifier.

Arguments

metaicvi::MetaICVIModule: metaicvi module to save with.

x::RealMatrix: features to train on.

y::IntegerVector: correct/over/under partition targets.

MetaICVI.Rocket.RocketKernel — Type

RocketKernel

Structure containing information about one rocket kernel.

MetaICVI.Rocket.RocketModule — Type

RocketModule

Structure containing a vector of rocket kernels.

MetaICVI.Rocket.RocketModule — Method

RocketModule(input_length::Integer, n_kernels::Integer)

Create a new RocketModule structure, requiring feature length and the number of kernels.

MetaICVI.Rocket.RocketModule — Method

RocketModule()

Default constructor for the RocketModule object.

MetaICVI.Rocket.apply_kernel — Method

apply_kernel(kernel::RocketKernel, x::RealVector)

Apply a single RocketModule kernel to the sequence x.

Arguments

kernel::RocketKernel: rocket kernel used for computing features.

x::RealVector: data sequence for computing rocket features.

MetaICVI.Rocket.apply_kernels — Method

apply_kernels(rocket::RocketModule, x::RealVector)

Run a vector of rocket kernels along a sequence x.

Arguments

rocket::RocketModule: rocket module containing many kernels for processing.

x::RealVector: data sequence for computing rocket features.

MetaICVI.Rocket.load_rocket — Function

load_rocket(filepath::String="rocket.jld2")

Load and return a rocket module with existing parameters from a .jld2 file.

Arguments

filepath::String: path to .jld2 containing rocket parameters. Defaults to rocket.jld2.

MetaICVI.Rocket.save_rocket — Function

save_rocket(rocket::RocketModule, filepath::String="rocket.jld2")

Save the rocket parameters to a .jld2 file.

Arguments

rocket::RocketModule: rocket module to save. filepath::String: path to .jld2 for saving rocket parameters. Defaults to rocket.jld2.